garbage-detection-main/requirements.txt
blinker==1.6.2
boto3==1.17.52
botocore==1.20.112
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.3.7
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0